!kick and !ban are now fully functional.
Added !kick and !ban commands to ScoobyBot, and they both check for permission before carrying out the feature. Had to make sure the bot had permission to kick/ban in server in order to work.

Next time I want to add a functionality so that the bot will check to see if itself has the permissions necessary to carry out kick/ban, along with clear.