Added !image command
Added an !image command that allows the user to query for an image. The bot then searches google and provides the URL to the first image result. Discord automatically converts image URL to an image in chat. I want to find a faster implementation, as it currently takes about 5 seconds to scrape an image and send the URL in chat.